+++ packages/core/src/flows/index.js
@@ -73,6 +73,7 @@ export {
 export { stakeAction } from './stakeAction.js';
 export { unstakeAction, claimRewardsAction } from './unstakeClaimActions.js';
 export { delegateAction, revokeDelegationAction } from './delegateRevokeActions.js';
+export { broadcastsForAddress } from './broadcastQueries.js';
 export {
     contractValidate,
     contractCheckCodeSize,